backend/requirements.txt
fastapi
uvicorn
scikit-learn
lightgbm
pandas
numpy
joblib
python-multipart
tld
python-whois
